HMS-1623 feat: add filter component
This change adds a composite component to enter the text to filter
for the server tab in the detail view.

We don't use this component (we decided to defer implementing it),
but because this component was already written, we decided to commit
it.  It is available for use at a later time.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>