fix(HMS-3165): container metadata files
It was observed some differences between the container generated from
the CI/CD platform system and the container generated locally from the
repository.

This change add several changes to try to be aligned with that changes
and generate the same or similar content to what is generated by the
platform.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>