feat(HMS-2992): add confirmation modal
During the domain registration wizard, when the
user click on cancel button, now a confirmation
modal dialog is showed up.

When the action is confirmed, a delete operation is
launched on the API; afterward the modal dialog is
closed and go to the main view.

When the action is not confirmed, the modal dialog
is closed with no additional action.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>