feat(HMS-2994): add delete confirmation modal
This change introduce the delete confirmation modal
dialog.

This dialog will show up when we click on the kebab
menu of the list domains view to delete the item, or
when we click on the kebab menu from the detail view
of the domain to delete the domain that is being
displayed.

In both cases, the deletion confirmation modal is
showed up, and when we press Delete button the
item is eventually deleted from the database.

Making click on the modal cross icon or cancel link
will dismiss the deletion confirmation modal with
no effect on the data stored.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>