fix(HMS-4059): notifications navigation issues
Notifications stopped working well when using wizard. The issue
seems to be that `useSelector` hook, that is used in `NotificationsPortal`,
stopped reacting to state changes and thus the notifications "stopped"
working well.

The situation looks to be gone after removing the lazy load of wizard pages.

This also improves a bit perceived responsivenes when already in the app
as the wizard pages are displayed right after navigating to them as there
is no load.

Removing the load should also be OK from code splitting perspective as
the pages are not big and thus could be part of the main app.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>