fix(HMS-4061): fix autoenrollment value
When the domain is registered on the backend, it is
created with autoenrollment disabled.

When the verification happens on the frontend, the
returned value is false, but we want to set by default
for the UI enabled. This change is not effective
until the user finish the wizard and click the Finish
button, and at that time the PATCH /domains/:domain_id
is sent and whatever decided the user (leave disabled,
or enabled) is then set on the backend side.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>