refactor(HMS-4079): build and lint fixes after PF5 update
`fec` build had issues with finding `PageSectionTypes`. To make
it working, it would probably be needed to update `COMPONENTS_CACHE`
in `tsc-transform-imports`. It was easier to replace with string.

`SampleComponent` was breaking build its props definition needed to
be more explicit.

Upgrade of `eslint-config-redhat-cloud-services` or related package
caused that eslint failed in most PF imports with
`rulesdir/forbid-pf-relative-imports` rule. Thus this rule was disabled
to keep the current change small.

hotReload defintion in fec config was changed as FEC warned that previous
way was deprecated.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>