refactor(HMS-4079): fix page padding/height
With upgrade to PF5, the size of content of idmsvc-frontend was
heigher than the space offered. I.e. scrollbars were displayed in
default page even when there was not enough content or in wizard,
the row with bottom buttons was not visible and required scrolling.

`<Page>` component is typically a high-level container which e.g.
contains also `Masthead` and `Navigation` but this is part of "chrome".
So in this app, all "page" is actually just the "page content" and
thus "PageGroup" is the suitable component.

This change fixes the size/padding issues so scrollbars are not
displayed when not needed.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>