fix(HMS-4253): do not change url when missing RBAC permissions
When accessing a page where a user doesn't have neccessary permissions
for the page, the user was previously redirected to a "NoPermissions"
page. This is changing it in a way that user is not longer redirected
but the NoPermission content is used for the current page instead of the
current content.

Thus it won't change page URL and allows to refresh the page to get the content
if users permissions change.

https://issues.redhat.com/browse/HMS-4253

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>