build: `npm run start-beta` target
When running `make run` or `npm run start`, developer has to manually
specify HCC environment, e.g. `stage`, and UI environment, e.g., `beta`.

This is "annoying" when the values are always the same, which atm are.
Thus introducing a `make run start-beta` which defines these two params
and thus it starts serving with single command.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>