feat(HMS-4265): skeleton as loading indicator
Improving the loading indicators UX with skeleton.

When loading RBAC or first load of domains, still show spinner as
it is not known what will be displayed.

When some domains were loaded, use skeleton table instead of spinner
when refreshing the domains table.

In details page, use skeleton as place holders for the form and
title.

Also fixes a bug in pagination, by removing offset, it then shows
correct Y in showing X - Y out of Y + adding a bit of margin between
pagination and the table.

https://issues.redhat.com/browse/HMS-4265

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>