chore(HMS-4521): ephemeral logging - debug level and source location
Debug level and source code location logging was set for idmsvc-backend
so that the template is more usable out of the box. The default was "warn".
Thus it is easier to investigate requests done to the API.

Source code location logging is now consistent with idmsvc-backend repo.

Also fixes incorrect GitHub namesapce.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>